requirements.txt
flask
pyyaml
colorama
wmi
py-cpuinfo
psutil
requests